Move db manage into own file and make util file.
