+++ indra/db/util.py
@@ -1,0 +1,89 @@
+from __future__ import absolute_import, print_function, unicode_literals
+from builtins import dict, str
+
+__all__ = ['get_defaults', 'get_primary_db']
+
+import os
+from os import path
+from .manage_db import DatabaseManager, IndraDatabaseError
+
+
+def get_defaults():
+    "Get the default database hosts provided in the specified `DEFAULTS_FILE`."
+    default_default_file = os.path.join(
+        os.path.dirname(__file__),
+        'defaults.txt'
+        )
+    env_key_dict = {'primary': 'INDRADBPRIMARY', 'test': 'INDRADBTEST'}
+    env = os.environ
+    available_keys = {k: v for k, v in env_key_dict.items() if v in env.keys()}
+    if not path.exists(default_default_file) and not available_keys:
+        raise IndraDatabaseError(
+            "Cannot find default file or environment vars."
+            )
+    elif path.exists(default_default_file):
+        with open(default_default_file, 'r') as f:
+            defaults_raw = f.read().splitlines()
+        defaults_dict = {}
+        for default_line in defaults_raw:
+            key, value = default_line.split('=')
+            defaults_dict[key.strip()] = value.strip()
+    else:
+        defaults_dict = {
+            purpose: env_val for purpose, my_env_key in env_key_dict.items()
+            for env_key, env_val in env.items() if my_env_key == env_key
+            }
+    return defaults_dict
+
+
+def get_primary_db(force_new=False):
+    """Get a DatabaseManager instance for the primary database host.
+
+    The primary database host is defined in the defaults.txt file, or in a file
+    given by the environment variable DEFAULTS_FILE. Alternatively, it may be
+    defined by the INDRADBPRIMARY environment variable. If none of the above
+    are specified, this function will raise an exception.
+
+    Note: by default, calling this function twice will return the same
+    `DatabaseManager` instance. In other words:
+
+    > db1 = get_primary_db()
+    > db2 = get_primary_db()
+    > db1 is db2
+    True
+
+    This means also that, for example `db1.select_one(db2.TextRef)` will work,
+    in the above context.
+
+    It is still recommended that when creating a script or function, or other
+    general application, you should not rely on this feature to get your access
+    to the database, as it can make substituting a different database host both
+    complicated and messy. Rather, a database instance should be explicitly
+    passed between different users as is done in the `by_gene_role_type`
+    function's call to `get_statements` in `indra.db.query_db_stmts`.
+
+    Parameters
+    ----------
+    force_new : bool
+        If true, a new instance will be created and returned, regardless of
+        whether there is an existing instance or not. Default is False, so that
+        if this function has been called before within the global scope, a the
+        instance that was first created will be returned.
+
+    Returns
+    -------
+    primary_db : DatabaseManager instance
+        An instance of the database manager that is attached to the primary
+        database.
+    """
+    defaults = get_defaults()
+    if 'primary' in defaults.keys():
+        primary_host = defaults['primary']
+    else:
+        raise IndraDatabaseError("No primary host available in defaults file.")
+
+    global __PRIMARY_DB
+    if __PRIMARY_DB is None or force_new:
+        __PRIMARY_DB = DatabaseManager(primary_host, label='primary')
+        __PRIMARY_DB.grab_session()
+    return __PRIMARY_DB
